Changed Current_Amount data type to "Bigint"/ long

+++ sl-server/src/main/java/com/github/jayreturns/slserver/user/domain/User.java
@@ -23,6 +23,6 @@ public class User {
     private String FriendlyName;
 
     @Column(name = "current_amount")
-    private Double CurrentAmount;
+    private long CurrentAmount;
 
 }
